배포 직전 전수 재조사에서 찾은 최적화 부채 29건 + 실기기 검증 지적 11건을 처리했다.
상세는 GL_MD/Perf_iOS_Optimization.md 6~7절 · Perf_Android_Optimization.md 6절.

■ 공유(GL_Shared) — 양 플랫폼 공통
지출 파생값 8중 순회 → 1패스(SpendingDerived 신설) · 가챠 정렬 키 선계산 + 통계·대시보드
순회 통합 · isSameMonth 가 yearMonthKey 를 쓰도록 · 타임존 캐시 우회 4곳 · Regex 상수화 12개 ·
게임 이름 인덱스 O(1) · 전월 대비 O(게임수×지출수) 제거 · 클라우드 push IO 이동

■ iOS
집계를 body 밖으로(인사이트·마이페이지·캘린더·게임정보) · DateMillis.comps 브리지 3회→1회 ·
공지 본문 이미지 캐싱 · 동시 요청 합치기 · 연간 리포트가 금액 수정에 반응하지 않던 버그

■ Android
전역 시머 클럭 게이팅(유휴 시 프레임 발행 0) · collectAsStateWithLifecycle 117곳 ·
사문화 로드인 코드 제거 · 보유 캐릭터 LazyColumn · Calendar 남용 · 마이페이지 파생 flow ·
탭 하단 여백 통일 · Baseline Profile 재생성(38,006 → 39,790 rules)

■ 유물 평가를 게임별 커뮤니티 표준으로
원신=아카샤 CV(치확×2+치피) / 스타레일·젠레스=유효 롤. 지표를 화면에 함께 밝힌다.
유효옵션 표기를 인게임 옵션명으로 통일하고, 젠레스 '이상 마스터리'가 원신 EM 으로
오분류돼 강조·점수에서 빠지던 버그를 고쳤다.

검증: commonTest 113 → 169건 · 양 플랫폼 Release 빌드 · iPhone 17 Pro Max / Galaxy S23
실기기 육안 검증 완료.

+++ GL_Shared/src/iosMain/kotlin/com/gatcha/log/storage/KeyValueStore.ios.kt
@@ -38,6 +38,13 @@ actual class KeyValueStore actual constructor(name: String) {
     // suiteName 으로 prefs 파일을 분리 (SharedPreferences 의 name 분리에 대응)
     private val defaults = NSUserDefaults(suiteName = name)
 
+    // ⚠️ 읽기를 objectForKey 한 번으로 줄이려고 `as? NSNumber` 로 바꿨다가 **되돌렸다**(2026-07-28).
+    //
+    // Kotlin/Native 가 `id` 로 받은 NSNumber 를 Kotlin NSNumber 로 주는지, 원시 타입으로 매핑하는지
+    // 실기기로 확인하지 않았다. 매핑이 다르면 캐스트가 조용히 null 이 되어 **모든 Int/Long/Boolean
+    // 설정이 기본값으로 읽힌다** — 예산이 0으로 읽히면 그대로 클라우드 스냅샷에 올라간다.
+    // 아끼는 건 조회 1회고 잃는 건 데이터다. 실측 근거가 생기면 그때 다시 본다.
+
     actual fun getString(key: String, default: String?): String? =
         (defaults.objectForKey(key) as? String) ?: default
 
@@ -49,7 +56,7 @@ actual class KeyValueStore actual constructor(name: String) {
     actual fun putBoolean(key: String, value: Boolean) = defaults.setBool(value, key)
 
     actual fun getLong(key: String, default: Long): Long =
-        if (defaults.objectForKey(key) != null) defaults.integerForKey(key).toLong() else default
+        if (defaults.objectForKey(key) != null) defaults.integerForKey(key) else default
 
     actual fun putLong(key: String, value: Long) = defaults.setInteger(value, key)
